# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VidVault AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /videos:
    get:
      operationId: listVideos
      tags: [videos]
      summary: List all videos for the current user
      parameters:
        - name: folderId
          in: query
          schema:
            type: string
            nullable: true
        - name: tagId
          in: query
          schema:
            type: string
            nullable: true
        - name: search
          in: query
          schema:
            type: string
            nullable: true
        - name: favorites
          in: query
          schema:
            type: boolean
            nullable: true
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of videos
          content:
            application/json:
              schema:
                type: object
                properties:
                  videos:
                    type: array
                    items:
                      $ref: "#/components/schemas/Video"
                  total:
                    type: integer
                required:
                  - videos
                  - total
    post:
      operationId: addVideo
      tags: [videos]
      summary: Add a new video by URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddVideoRequest"
      responses:
        "201":
          description: Video added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        "400":
          description: Invalid URL or video not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /videos/{videoId}:
    get:
      operationId: getVideo
      tags: [videos]
      summary: Get a single video
      parameters:
        - name: videoId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Video details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoDetail"
        "404":
          description: Video not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateVideo
      tags: [videos]
      summary: Update a video
      parameters:
        - name: videoId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVideoRequest"
      responses:
        "200":
          description: Updated video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
    delete:
      operationId: deleteVideo
      tags: [videos]
      summary: Delete a video
      parameters:
        - name: videoId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Video deleted
  /videos/{videoId}/favorite:
    post:
      operationId: toggleFavorite
      tags: [videos]
      summary: Toggle favorite status
      parameters:
        - name: videoId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated favorite status
          content:
            application/json:
              schema:
                type: object
                properties:
                  isFavorite:
                    type: boolean
                required:
                  - isFavorite
...
components:
  schemas:
...
    Video:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        url:
          type: string
        title:
          type: string
        thumbnail:
          type: string
          nullable: true
        duration:
          type: string
          nullable: true
        channelName:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        folderId:
          type: string
          nullable: true
        folderName:
          type: string
          nullable: true
        isFavorite:
          type: boolean
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        viewCount:
          type: integer
          nullable: true
        publishedAt:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - url
        - title
        - isFavorite
        - tags
        - createdAt
        - updatedAt
    VideoDetail:
      allOf:
        - $ref: "#/components/schemas/Video"
        - type: object
          properties:
            notes:
              type: array
              items:
                $ref: "#/components/schemas/Note"
            aiOutputs:
              type: array
              items:
                $ref: "#/components/schemas/AiOutput"
          required:
            - notes
            - aiOutputs
    AddVideoRequest:
      type: object
      properties:
        url:
          type: string
        folderId:
          type: string
          nullable: true
      required:
        - url
    UpdateVideoRequest:
      type: object
      properties:
        folderId:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
...
    Tag:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
          nullable: true
      required:
        - id
        - name
...
    Note:
      type: object
      properties:
        id:
          type: string
        videoId:
          type: string
        userId:
          type: string
        content:
          type: string
        timestamp:
          type: integer
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - videoId
        - userId
        - content
        - createdAt
        - updatedAt
...
    AiOutputType:
      type: string
      enum:
        - summary
        - notes
        - ppt_outline
        - mcq
        - flashcards
        - blog_article
        - key_insights
    AiOutput:
      type: object
      properties:
        id:
          type: string
        videoId:
          type: string
        type:
          $ref: "#/components/schemas/AiOutputType"
        content:
          type: string
        createdAt:
          type: string
      required:
        - id
        - videoId
        - type
        - content
        - createdAt
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
          nullable: true
      required:
        - error