# lib/api-zod/src/generated/types/updateVideoRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VidVault AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateVideoRequest {
  folderId?: string | null;
  title?: string | null;
}