# lib/api-zod/src/generated/types/video.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VidVault AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Tag } from "./tag";

export interface Video {
  id: string;
  userId: string;
  url: string;
  title: string;
  thumbnail?: string | null;
  duration?: string | null;
  channelName?: string | null;
  description?: string | null;
  folderId?: string | null;
  folderName?: string | null;
  isFavorite: boolean;
  tags: Tag[];
  viewCount?: number | null;
  publishedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}