# lib/api-zod/src/generated/types/videoDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VidVault AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiOutput } from "./aiOutput";
import type { Note } from "./note";
import type { Video } from "./video";

export type VideoDetail = Video & {
  notes: Note[];
  aiOutputs: AiOutput[];
};